+++ packages/react-native-reanimated/src/ReanimatedModule/NativeReanimated.ts
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/no-empty-function */
 'use strict';
 
 import type { SerializableRef, WorkletFunction } from 'react-native-worklets';
@@ -64,16 +63,10 @@ class NativeReanimatedModule implements IReanimatedModule {
     }
     global._REANIMATED_VERSION_JS = jsVersion;
 
-    if (ReanimatedTurboModule) {
-      const status = installTurboModule();
-      if (!status) {
-        // This path means that React Native has failed on reload.
-        // We don't want to throw any errors to not mislead the users
-        // that the problem is related to Reanimated.
-        // We install a DummyReanimatedModuleProxy instead.
-        this.#reanimatedModuleProxy = new DummyReanimatedModuleProxy();
-        return;
-      }
+    if (ReanimatedTurboModule && !installTurboModule()) {
+      throw new Error(
+        '[Reanimated] Failed to install the native module because React Native has no active surface on its discarded instance. This happens when the app is reloaded while a previous reload is still in progress, for example during an OTA update immediately upon launch. No action is required; the app will continue to function as usual.'
+      );
     }
 
     if (global.__reanimatedModuleProxy === undefined) {
@@ -261,52 +254,6 @@ See https://docs.swmansion.com/react-native-reanimated/docs/guides/troubleshooti
   unregisterPseudoStyles(viewTag: number) {
     this.#reanimatedModuleProxy.unregisterPseudoStyles(viewTag);
   }
-}
-
-class DummyReanimatedModuleProxy implements ReanimatedModuleProxy {
-  configureLayoutAnimationBatch(): void {}
-  setShouldAnimateExitingForTag(): void {}
-  getStaticFeatureFlag(): boolean {
-    return false;
-  }
-  setDynamicFeatureFlag(): void {}
-  subscribeForKeyboardEvents(): number {
-    return -1;
-  }
-
-  unsubscribeFromKeyboardEvents(): void {}
-  setViewStyle(): void {}
-  setCSSEventHandler(): void {}
-  markNodeAsRemovable(): void {}
-  unmarkNodeAsRemovable(): void {}
-  registerCSSKeyframes(): void {}
-  unregisterCSSKeyframes(): void {}
-  applyCSSAnimations(): void {}
-  registerCSSAnimations(): void {}
-  updateCSSAnimations(): void {}
-  unregisterCSSAnimations(): void {}
-  runCSSTransition(): void {}
-  unregisterCSSTransition(): void {}
-  registerSensor(): number {
-    return -1;
-  }
-
-  unregisterSensor(): void {}
-  registerEventHandler(): number {
-    return -1;
-  }
-
-  unregisterEventHandler(): void {}
-  getViewProp() {
-    return null!;
-  }
-
-  getSettledUpdates(): SettledUpdate[] {
-    return [];
-  }
-
-  registerPseudoStyles(): void {}
-  unregisterPseudoStyles(): void {}
 }
 
 function installTurboModule() {